Move PEC to an enterprise-kernel module
Since PEC is not related to HA, it makes sense for it to be in its
own module rather than in the HA module.
