+++ enterprise/ha/pom.xml
@@ -75,7 +75,7 @@
   <dependencies>
     <dependency>
       <groupId>org.neo4j</groupId>
-      <artifactId>neo4j-kernel</artifactId>
+      <artifactId>neo4j-enterprise-kernel</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
