+++ enterprise/kernel/LICENSES.txt
@@ -9,7 +9,6 @@ Apache Software License, Version 2.0
   opencsv
   parboiled-core
   parboiled-scala
-  The Netty Project
 ------------------------------------------------------------------------------
 
                                  Apache License
