+++ enterprise/kernel/src/test/java/org/neo4j/graphdb/factory/EnterpriseDatabaseRule.java
