--- enterprise/pec-tests/NOTICE.txt
@@ -1,45 +1,0 @@
-Neo4j
-Copyright © 2002-2015 Network Engine for Objects in Lund AB (referred to
-in this notice as "Neo Technology")
-   [http://neotechnology.com]
-
-This product includes software ("Software") developed by Neo Technology.
-
-The copyright in the bundled Neo4j graph database (including the
-Software) is owned by Neo Technology. The Software developed and owned
-by Neo Technology is licensed under the GNU AFFERO GENERAL PUBLIC
-LICENSE Version 3 (http://www.fsf.org/licensing/licenses/agpl-3.0.html)
-("AGPL") to all third parties and that license, as required by the AGPL,
-is included in the LICENSE.txt file.
-
-However, if you have executed an End User Software License and Services
-Agreement or an OEM Software License and Support Services Agreement, or
-another commercial license agreement with Neo Technology or one of its
-affiliates (each, a "Commercial Agreement"), the terms of the license in
-such Commercial Agreement will supersede the AGPL and you may use the
-software solely pursuant to the terms of the relevant Commercial
-Agreement.
-
-Full license texts are found in LICENSES.txt.
-
-Third-party licenses
---------------------
-
-Apache Software License, Version 2.0
-  Apache Commons Lang
-  ConcurrentLinkedHashMap
-  Jackson
-  Lucene Core
-  opencsv
-  parboiled-core
-  parboiled-scala
-  The Netty Project
-
-BSD - Scala License
-  Scala Library
-
-BSD License
-  ASM Core
-  Scala Compiler
-  scala-parser-combinators
-
