+++ enterprise/pom.xml
@@ -33,11 +33,11 @@
     <module>com</module>
     <module>cluster</module>
     <module>backup</module>
+    <module>kernel</module>
     <module>ha</module>
     <module>neo4j-enterprise</module>
     <module>server-enterprise</module>
     <module>enterprise-performance-tests</module>
-    <module>pec-tests</module>
   </modules>
 
   <licenses>
