+++ manual/cypher/cypher-docs/pom.xml
@@ -116,13 +116,13 @@
     </dependency>
     <dependency>
       <groupId>org.neo4j</groupId>
-      <artifactId>neo4j-ha</artifactId>
+      <artifactId>neo4j-enterprise-kernel</artifactId>
       <version>${project.version}</version>
       <scope>test</scope>
     </dependency>
     <dependency>
       <groupId>org.neo4j</groupId>
-      <artifactId>neo4j-ha</artifactId>
+      <artifactId>neo4j-enterprise-kernel</artifactId>
       <version>${project.version}</version>
       <type>test-jar</type>
       <scope>test</scope>
