+++ manual/cypher/refcard-tests/pom.xml
@@ -127,13 +127,13 @@
     </dependency>
     <dependency>
       <groupId>org.neo4j</groupId>
-      <artifactId>neo4j-ha</artifactId>
+      <artifactId>neo4j-enterprise-kernel</artifactId>
       <version>${project.version}</version>
       <scope>test</scope>
     </dependency>
     <dependency>
       <groupId>org.neo4j</groupId>
-      <artifactId>neo4j-ha</artifactId>
+      <artifactId>neo4j-enterprise-kernel</artifactId>
       <version>${project.version}</version>
       <type>test-jar</type>
       <scope>test</scope>
